Download all working

+++ invoice_system_management/invoice/models.py
@@ -15,7 +15,7 @@ class Product(models.Model):
 class Customer(models.Model):
     customer_name = models.CharField(max_length=255)
     customer_contact = models.CharField(max_length=255)
-    customer_amount = models.IntegerField(default=0)
+    customer_amount = models.FloatField(default=0)
     customer_is_delete = models.BooleanField(default=False)
 
     def __str__(self):
